Remove unnecessary type conversions.

+++ grains/grains.go
@@ -12,10 +12,10 @@ func Square(n int) (uint64, error) {
 		return 0, errors.New("invalid square")
 	}
 
-	return uint64(1 << uint(n-1)), nil
+	return 1 << uint(n-1), nil
 }
 
 // Total calculates the total number of grains for the entire chessboard
 func Total() uint64 {
-	return uint64((1 << totalSquares) - 1)
+	return (1 << totalSquares) - 1
 }
